requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mysql-connector-python==8.0.22
oauthlib==3.1.0
pip==20.3.3
protobuf==3.14.0
PySocks==1.7.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
setuptools==51.1.1
six==1.15.0
soupsieve==2.1
tweepy==3.10.0
urllib3==1.26.2